Making page look and feel cleaner + putting phase 2 in their own folders.
Added the DUO logo where needed, as well as making the slide toggles better and making it where things are hidden on click. Also making it to where when you logout somethings don't slide into the login page. As well as fitting all the phase 2 stuff into its own folder so the folders we have look cleaner.